refactor(users): change repositories folder structure
